requirements.txt
pip==20.0.2
Django==3.0.4
django-bootstrap3==12.0.3
django-crispy-forms==1.9.0
coverage==5.0.4
plotly==4.5.4
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.2.0
